+++ core/CloudXCore.xcframework/ios-arm64_x86_64-simulator/CloudXCore.framework/Headers/CLXAdNetwork+Internal.h
@@ -27,6 +27,8 @@ extern CLXAdNetwork const CLXAdNetworkVungle;
 extern CLXAdNetwork const CLXAdNetworkInMobi;
 extern CLXAdNetwork const CLXAdNetworkUnityAds;
 extern CLXAdNetwork const CLXAdNetworkMagnite;
+extern CLXAdNetwork const CLXAdNetworkMoloco;
+extern CLXAdNetwork const CLXAdNetworkVerve;
 
 /// All known ad networks.
 extern NSArray<CLXAdNetwork> *CLXAllAdNetworks(void);